pyproject.toml
[project]
name = "monomorph"
version = "0.2.0"
description = "MonoMorph: a LLM-powered agentic tool for refactoring monolithic codebases into microservices."
requires-python = ">=3.12"
dependencies = [
    "colorama==0.4.6",
    "colorlog>=6.10.1",
    "decparsing>=1.6.1",
    "docker==7.1.0",
    "fuzzywuzzy===0.18.0",
    "grpcio==1.71.0",
    "grpcio-health-checking==1.71.0",
    "jinja2==3.1.4",
    "langchain==0.3.21",
    "langchain-community>=0.3.20",
    "langchain-google-genai==2.1.6",
    "langchain-google-vertexai==2.0.17",
    "langchain-openai==0.3.11",
    "langgraph==0.3.24",
    "langmem==0.0.28",
    "numpy==2.2.6",
    "openai==1.78.1",
    "pandas==2.2.3",
    "protobuf==5.29.4",
    "pydantic==2.11.7",
    "python-dotenv==1.0.1",
    "typing-extensions==4.14.0",
]

[dependency-groups]
dev = [
    "pydeps>=3.0.1",
]
